Fixed Issue 116 Changed volumes from double to float, because in the end we need float, because that's what the Android API wants. See http://developer.android.com/reference/android/media/MediaPlayer.html#setVolume(float, float)
Also, I was unable to think of a test for this, because while we can _set_ the volume of individual MediaPlayers, there is currently no way of getting it, besides remembering it.

Update issue 116
Owner: aried3r@gmail.com
I fixed this issue, note however that the volume is a global one amongst all media players. When setting the volume to 100%, _all_ sounds in _all_ sprites will have 100%.