Update Issue 211 Added Japanese translation for category_lego_nxt. Renamed Consts to Constants and updated the references.
As for the other languages which are missing some strings: We could kindly ask our GSoC students to translate some of them. But if the strings are not existent, the default - English - ones will be used. As our Russian/Romanian userbase is currently quite small, I suggest we neglect it for the moment, maybe even wait until we have a RC for v7 of Catroid, so we have a finalized version of strings.xml instead of adding, removing, correcting them every few commits.

If you volunteer to translate the Japanese ones bit by bit, we can add Japanese to the unit tests, which look for untranslated strings (currently only active for German and English, as this is what 90%+ of our developers speak and can correct/add by themselves).