Added new strings, deleted old ones.
Removed the English and German strings that are supposedly used in
MainMenuActivty, but weren't from what I can see. Then added the new
ones with a prefix 'main_menu_'.

Removed now obsolete icons, replaced them with <shape> drawables. To be
found in /res/drawable. They are all the same but already have the
correct names, so once the icons arrive, they can be just dropped in.