Refactored test in an attempt to escape OOM Error
While unable to find our memory leak on orientation change so far, I
adopted some tests that were especially prone to leak memory, leading to
OOM errors. Of course this is not a real solution but a workaround. I'll
keep on digging in the Heap and Allocation analysis tools to find it.

Even if we, in the upcoming release, lock Catroid to portrait mode, it
would be better to detect the leak anyway.