Removed some static imports
Removed the static imports from our Constants class in the .xml package.
I used static imports for the XML constants within our .xml package,
because in there it - in my opinion, anyway - makes sense and does not
decrease readibility but on the contrary remove some clutter.
Within the .xml package, you expect constants to be related to the code,
which they are. Using static imports for other constants breaks that
concept, I removed them again.