Fixed Issue #118
The Nullpointer occured when the test silently failed and was executed
again due the @FlakyTest annotation. Since the Sprite was a member and
only initialized once, the already running StartScript from the first
try was still running in the background, still calling getScript() in
the LoopEndBrick, which returned null, because all Scripts had been
removed via removeAllScripts() and replaced with a new one, which did
not contain the LoopEndBrick anymore.