Fixed crash when no SD Card is present
* Check for SD Card improved. We now not only look for mounted media but
  also make sure it's not READ_ONLY. The new function name reflects
  that.
* Added the check in MainMenuActivities onCreate() so we display an
  error message and quit gracefully instead of crashing.

I tried to write a test for this using PowerMock + Mockito but did not
succeed. In fact, I did not get to the point where I could even compile
with these libraries added. Android Support was added to Mockito in
1.9.5 but it seems there is no ready-to-use PowerMock package for
Android. I encountered various errors while trying to make a runnable
package.
The other approach would be to rewrite the functions with e.g.
Dependency Injection so we can insert e.g. a MockEnviroment to the Utils
function or similar approaches.
In the end, I decided to not do test-first in this case so this critical
issue can get closed and find a way to test it later on.