Removed Native App project until fixed in branch
nativeAppBuilding project doesn't work anymore. Talked with Severin and
Wolfgang, agreed to remove the project from the master branch until we
re-add it again once it's finished and working again.

What's NOT been done:
Remove all traces of NativeApp from Java projects including tests. I'm
not yet sure what to remove there, but it doesn't hurt because the code
either is never executed and the tests, if any, work too.

This step was taken as a precauting because soon the GSoC will start
again and we shouldn't have any projects in the master branch that
cannot be successfully run.