Changed the manual solution to use the Android Quantity Strings.
Implemented a Utils function that handles the double -> integer plurals
problem and can be adapted, and works for at least english/ german in
the upcoming release. 

Also added the distinction to the getPrototypeView() functions to
prevent the template bricks displaying a different plurality than they
get by default when inserted.