TextFields not using space on larger devices
fixes #570
I created a new ViewGroup called BrickLayout which replaces
LinearLayout(s) inside the bricks. It follows the specifications given
by the UX designers:
 1. text fields all have the same maximum size.
 2. text fields use the maximum amount of space possible
 3. on small screens, use line breaks to maximize text field width. On large
    screens, eliminate as many line breaks as possible.
 4. the line breaks are static, I.E., determined at the time the brick is
    shown for the first time, never reshuffled as the text field sizes
    change.

I ended up changing all the EditTexts on bricks to TextViews to solve
the elipsizing problem, and as a result, I updated the
EditTextAlignmentTest to reflect those changes.