Bottom bar hidden on action mode
Fixes #641
The BottomBar is now hidden everywhere where it is not intended to be,
like renaming/deleting objects (instead of disabling the buttons in the
Bottombar).
Also the BottomBar.java class has changed. Now there are methods for
showing and hiding the BottomBar and the Add+Play - Button.
The Bottombar separator is now handled automatically by BottomBar.java
(was: setting separator visibility manually where it was needed).

Jenkins Testrun(s):
https://jenkins.catrob.at/job/Catroid-Multi-Job-Custom-Branch/355/

Ran the failed ui test separatly 10 times:
https://jenkins.catrob.at/view/Catroid/job/Catroid-single-UI-emulator/183/