increased version number - v.0.9.13
Like 0.9.11 minus the vibrate and flash bricks

Catroid 0.9.12 unnecessarily increased the language version to 0.93.
This reverts it to 0.92.

Conflicts:
	catroid/src/org/catrobat/catroid/ui/fragment/CategoryBricksFactory.java

revert language update of projects

fix some loading issues

Conflicts:
	catroid/src/org/catrobat/catroid/ProjectManager.java
	catroid/src/org/catrobat/catroid/io/StorageHandler.java

lang version 0.93

check the CategoryBricksFactory Class if Sprite is needed in Bricks!!!
Conflicts:
	catroid/src/org/catrobat/catroid/ui/fragment/CategoryBricksFactory.java

Like 0.9.11 minus the vibrate and flash bricks

Catroid 0.9.12 unnecessarily increased the language version to 0.93.
This reverts it to 0.92.