CAT-1374 Add checkstyle rule EmptyCatchBlock
To allow empty catch blocks, the exception variable name has to be
'expected', 'ignore' or 'ok'. This confirms with the google style guide
and can be useful for tests where an exception should be thrown.

The PMD rule EmptyCatchBlock has been excluded to allow empty catch
blocks in the cases explained above. Also the unnecessary code in the
expected catch blocks has been removed.

http://google-styleguide.googlecode.com/svn/trunk/javaguide.html#s6.2-caught-exceptions
https://jira.catrob.at/browse/CAT-1374