CAT-1438 Fixed testAddNewLook in SetLookBrickTest:
This test failed a lot, because we were not able to do a real click
on the Paintroid Option. We have to use the MockActivity and initiate the
click with startActivityForResult, which is not the same as clicking on
the Option. Due to this false click, some Code was not triggered, and therefore
the App didn't send the Broadcast for a new LookFile. I fixed this
issue by clicking on the Spinner, where we then can see the new File.