CAT-1636 Order XStream XML fields by annotation
The new XStreamFieldKeyOrder annotation provides an easier way to define
a custom field order in the XML file.

Only serialized (!static and !transient) fields have to be specified in
the annotation now. If not specified, the CatroidFieldKeySorter will
throw an exception when saving a project. If there is an alias defined
for the field, the alias has to be used in the annotation.

There are two reasons why I decided to throw an exception
- It's identical to the previous behavior, just the exception type
  differs.
- A simple log message will easily be overseen.

http://xstream.10960.n7.nabble.com/Is-there-a-way-to-set-the-order-of-Xstream-serialization-td8300.html#a8464