Add CatroidFieldKeySorterTest
Tests the general case of the CatroidFieldKeySorter where fields are
sorted alphabetically by class hierarchy. The current implementation
works like following. All fields in the topmost base class are sorted
alphabetically and added to the list order, then the second topmost base
class fields are sorted and added to the list and so on.