ARDUINO-21 WhenRaspiBrick: interrupt based Scripts & settings
Implement a script that is triggered by a pin state change
on a connected raspberry pi.
WhenRaspiPinChangedBrick utilizes PocketCode's broadcast message
system. The raspi broadcast message start with a special identifier
Constants.RASPI_BROADCAST_PREFIX and are not listed within the original
BroadcastReceiverBrick.

the version of the raspberryPi can be selected in the sections.
only available pins can be selected via the pin spinner.