CAT-1078 Fix transparent passepartout issue
Looks and the passepartout share the same batch during drawing the
stage. Each Look sets the color and alpha value of the batch according
to its settings before drawing itself [1]. The passepartout however
doesn't and therefor uses the alpha value set by the previous Look. So,
when the previous Look sets the alpha value to 0.5, the passepartout
will also be drawn with an alpha value of 0.5 and the background and
other images are shining through it.

By explicitely setting the color of the batch to black and the alpha
value to 1 before drawing the passepartout, the passepartout will remain
black.

[1] https://github.com/libgdx/libgdx/blob/97d6e8d/gdx/src/com/badlogic/gdx/scenes/scene2d/ui/Image.java#L119

https://jira.catrob.at/browse/CAT-1078