CAT-1779 Add Collisiondetection of Looks
The check for collision between looks can be achieved through the
FormulaEditor. If two looks are selected, two threads are started to
calculate the corresponding collision polygons. The resulting data
is written to the metadata of the png file, so calculations for
a specific look have to be done only once per project. If it is not
a png file, the simple outlining hitbox is used as the collision
polygon.
These polygons are the outlines of the picture used by the look; they
are not convex and if there is a hole or something similar, it is also
detected as a polygon.

When checking the collision on stage, there is a broad phase and a
narrow phase.

- Broad phase: Check if the hitboxes even overlap.

- Narrow phase: Check for intersection of any of the polygons vertices.
  If this doesn't yield a positive result, checks are done, if one polygon
  is contained in the other one (which is specified as a collision).