HOTFIX Fix conversion problem w. special chars in program name
Programs of an older catrobat language version are converted to the new one.
For this purpose, it is necessary to create a DefaultScene.
The migration did not copy images and sounds due to double-escaping the path.
This led that "Foo:bar" was first escaped to "Foo%3ABar" and then in the second
step, the '%' was (erroneously) escaped again. Because of that, the file
path was incorrect. This commit fixes that issue.g