CAT-FIX Fix LookData reference in Bricks for clones
It is important that the LookDatas are cloned before the Bricks are
cloned. Otherwise the SetLookBrick LookData would refer to an origin
LookData instead of a cloned one.

It is also important that the Look is created and cloned after the
Bricks are cloned. Otherwise always a Look instance is created instead
of a Look or PhysicsLook.

To fix this problem, first, the LookDatas are cloned, then the Bricks,
then the Look.

https://github.com/Catrobat/Catroid/pull/1973#issuecomment-256002440