CAT-1384 Copy all Look/PhysicsLook/PhysicsObject values when cloning
When cloning a Sprite in the Stage or in the IDE, values like position,
direction, mass, etc have to be copied as well.

* Clone missing value size in Look. All other values are already cloned.
  [1][2]
* Clone all values in PhysicsLook and PhysicsObject

The LookData can't be cloned *in* the Look because each Look has to
refer to its Sprites LookData. It is already correctly cloned in the
Sprite. [2]

The z-index can't be cloned.

In the IDE, there is no Stage. Setting the z-index of an Look which
hasn't been added to a Stage has no effect and calling getZIndex will
always return -1.

In the Stage, it isn't possible that two Looks have the same z-index.

[1] https://github.com/Catrobat/Catroid/pull/1972
[2] https://github.com/Catrobat/Catroid/pull/1980
https://jira.catrob.at/browse/CAT-1384