CAT-REF Remove paused member from Script and Sprite
The paused member was used in the old implementation when we handled the
action execution by threads. A long time ago we switched to libGDX
actions. So there is no need for the member anymore.

The only case where the member is used, is in Look.doTouchDown [1].
However, when pausing the stage, a grey overlay covers the whole stage.
All touch events are captured by this overlay and no stage touch events
are fired.

It's important that no touch events are fired when the stage is paused
for Sprites and TouchUtil.

[1] https://github.com/Catrobat/Catroid/blob/2ce3242fda5f8c020757a9e31fd6d78f111822c5/catroid/src/main/java/org/catrobat/catroid/content/Look.java#L164-L166