CAT-FIX XStream sort by annotation with derived classes
Derived classes do not inherit annotations and therefore will always be
sorted alphabetically. The CatroidFieldKeySorter fixes this problem by
searching for the XStreamFieldKeyOrder annotation in the class hierarchy
now. If a derived class adds new members, which aren't specified in a
base class, a XStreamMissingSerializableFieldException is thrown.

This fix currently addresses only the Sprite and its subclasses. The
Sprite specifies a custom field order but the subclasses don't and
therefore the fields were always sorted alphabetically. Now, each
Subclass looks for the XStreamFieldKeyOrder annotation in a base class,
the Sprite, and sorts it accordingly.

The @Inherited would've also solved this problem but is easily forgotten
and more error-prone.