CAT-FIX Use getCurrentScene in InternFormulaParser
Parsing only happens when editing a Formula through the
FormulaEditorFragment in the UI. So the InternFormulaParser uses
getCurrentScene only for now on.

Normally you can use getCurrentScene or getSceneToPlay in the UI. Both
are always refering to the same Scene Object when in the UI. However,
for correctness getCurrentScene is used because it represents the
currently opened Scene.