CAT-FIX Adapt checkstyle license regex check for XML files
Checkstyle forced to use the start comment tokens in XML files on its
own line.

<?xml version="1.0" encoding="utf-8"?>
<!--

Creating new files with Android Studio also adds the comment token on
its own line. However, when reformatting the XML, the comment token will
be moved to the first line.

<?xml version="1.0" encoding="utf-8"?><!--

Since this is a bug of Android Studio and there is no fix for it,
Checkstyle allows both versions for now on.

https://github.com/Catrobat/Catroid/pull/2152#discussion_r106465282