CAT-1152 Scale down screenshots to 480x480 when uploading a project.
Made ProjectAndScreenshotLoader's method to load a screenshot file
accessible.
When uploading a project, a list of scene names is added to the upload intent, as we need the scene names in order to access the screenshots.
ProjectUploadService loops through all given scene names and scales screenshots down to 480x480.
Overloaded ImageEditing method to scale with width and height in addition to a scaling factor.
Changed getScaledBitmapFromPath(...) to use the justScaleDown boolean flag (instead of doing nothing with it).
Adjusted the preparation of reuploadIntent to include the scene names in case a reupload hast to be done.
Rewrote if as switch statement in ProjectUploadService for better readability.

ImageEditingTest: Edited test to test for onlyScaleDown flag

ImageEditing: Made if statement in getScaledBitmapFromPath more readable; fixed a bug in getScaledBitmapFromPath where height was assigned to width and the other way around (only noticable since onlyScaleDown is implemented); split up getScaledBitmapFromPath into two methods for better reusability; removed unused getScaledBitmapOfLoadedBitmap(...) and getImageDimensionsForLoadedImage(...)

UtilFile: Some cosmetic refactoring

ProjectUploadService: Handled FileNotFoundException so that the upload is not stopped when scaling down a screenshot fails

ProjectAndSceneScreenshotLoader: Refactored getScreenshotFile to be a bit shorter