CAT-2353 Added removeUnusedBroadcastMessages() to project so that messages used by none of the scenes are removed
Added addUsedMessagesToList(...) to Scene to add all used messages to a given list.

Replaced call to scene.removeUnusedBroadcastMessages() with call to project.removeUnusedBroadcastMessages() in MergeTask and moved the new call to a different position (where the merged project already has all of its scenes).

Replaced call to scene.removeUnusedBroadcastMessages() with call to project.removeUnusedBroadcastMessages() in ScriptFragment.

Fixed formatting