CAT-2400 Automatize the creation of independent APKs.
Independent APKs are allowed to be installed without affecting the
current Pocket Code installation.
This is useful for testers to have different versions installed in parallel.
Automating this process (https://confluence.catrob.at/x/KoGT) makes it more
accessible and allows Jenkins to provide such independent APKs.

Usage
-----
./gradlew -Pindependent assembleDebug

The created catroid-catroid-debug.apk file has an internal application name
that depends on the current date and time.
The user displayed label is of the form "Code MMdd HH:mm" in UTC.

Alternatively you can provide an application name yourself:
./gradlew -Pindependent="Code Design #2" assembleDebug

Implementation
--------------
The application name and the package name are stored as placeholders in
android manifest and also in the google-services-template.json.
Moreover the test application id is not set manually anymore, since it
is created automatically by the android gradle plugin.