CAT-2145 remove range check for PWM values
This removes an inconsistency in the code and creates compatibility
with Arduino's value range.
All range checks are now left to Firmata and Arduino. Hence the
default range of [0-255] should be usable now, with a maximum range
of [0-16383] supported by Firmata.
Further details can be found in the JIRA discussion of this issue.