CAT-FIX Set CRASHLYTICS flag to True in debug mode
Earlier the flag CRASHLYTICS_CRASH_REPORT_ENABLED was set to false in
debug mode. In SettingsActivity the checkbox button (which
enables/disables the Crashlytics crash reporting) is either visible or
hidden depending upon the flag value (visible if true and hidden if
false).

As the checkbox was hidden (hidden for false), it resulted in failure of
different espresso test cases where checkbox was required to perform the
UI test. Moreover, espresso testrun is green because the auto-generated
AllEspressoTestsSuite.java file isn't up-to-date and doesn't contain the
SettingsActivityTest.