Use internal storage for standalone builds
This commit adds additional methods to StorageHandler, such as
getRootDirectory(), setRootDirectory() that allow setting the storage
location at runtime. This is necessary because we cannot obtain the path
to internal storage at compile time.
The application code was modified so that we do not use the root directory
defined in Constants.java directly any more, instead we call the
getRootDirectory() method to get the path.
The directory structure is now not created during initialization of the
StorageHandler, but lazily when we first call getRootDirectory().
This is necessary so that we can change the root directory before first use.

As a result of this change, there is no interference between
standalone apps and the PocketCode application.