JENKINS-208: Allow build steps to run on a local machine
Add buildScripts/build_*-scripts which allows the
developer to run the same build steps which run
on Jenkins on the local machine.

So the logic of the single build-steps is moved
to a separate script per step.

The scripts take care of creating and starting
the emulator. The AVD to use can be configured
via the buildScripts/emulator_config.ini-file.

Import the upstream jenkins-android-helper used to
handle the SDK installation and the emulator handling.