CAT-2732: Testing: Disable animations via gradle
For UI-Tests the animations shall be disabled, currently
this was done programmatically, which was broken for at
least some tests.
Now the animations are disabled before the testrun via
gradle. PocketMusicActivityTest relies on Animations, so
this single test-class enables animations before the
testrun and resets them to the old value afterwards.