CAT-FIX Fix crash when not connected to a network
When the device is not connected to a network,
ConnectivityManager.getActiveNetworkInfo will return null which will
result in a NullPointerException. Added null checks to prevent this
crash.

There is a small behavior change when an error occurs in the WebView.
Previously, the additional info '…Check your Internet connection.' was
only shown if the device is not connected to a network. Now the info is
also shown if the device is connecting to a network. IMO and in this
case, the difference is almost non-existent. Also, if the device is
connecting to a network, it doesn't mean it will get a connection. So
showing the additional info in both cases is fine.