CATROID-37 crashes when restoring fragments after activity was destroyed
Can be reproduced by setting Don't keep activities in developer options.
Problem is that some fragments have listeners or members that link back to
its activity or other views. This isnt recoverable when activity is killed
due to low memory or long inactivity. When app is left in this state it
will crash badly on reopening.
To fix this for now this PR prevents those fragments to be restored when
activity is recovering from such a killed state. (eg. formula editor open
- activity killed by OS - reopen activity is back in script view)