CATROID-46 SetLookByIndexAction created with wrong sprite reference
The SetLookByIndexAction was created by calling
`ProjectManager.getInstance().getCurrentSprite()` which will always
return the sprite currently open/editing in the IDE. This lead to
crashes or unexpected behavior when run.

Fixed by using the passed sprite parameter instead when calling
addActionToSequence. Reverting the changes made to addActionToSequence
(SetLookByIndexBrick and SetBackgroundByIndexBrick) in 33700038a.