CATROID-101 Remove paramters to not clash with the job dsl.
The job dsl has to create paramters for the git repository/branch.
Yet these two paramters clear all other paramters, also those defined in the Jenkinsfile.
Since the job dsl is run once per day it could result in users having
to start the ManualTests job twice.
Once to create the necessary paramters and a second time to fill in these paramters.

To avoid such inconvenience it is better to manage the paramters soley
via the job dsl.