JENKINS-278 Ensure the Jenkinsfile can also be used in a Pipeline job.
BRANCH_NAME is only specified in multi branch pipeline jobs.
It is null in regular pipeline jobs.
Thus adapt the Jenkinsfile so that it works with a null BRANCH_NAME.

That way the Jenkinsfile can be used in regular pipeline jobs, which is
useful for the development of the Jenkinsfile.