JENKINS-285 Integrate optional building of all Catroid flavors as parameter.
Before there was the Catroid-IndependentAPKs to build all flavors and archive the APKs.
That job needed a branch and repository to work.
Yet the common use-case to retrieve the APKs for a pull request was surprisingly hard.

Thus with these changes it is sufficient to manually start the job for the PR,
or any other Catroid branch for that matter, and to select the BUILD_ALL_FLAVORS option.