CATROID-224 ListWithoutDuplicates Implementation for RV
RecyclerView-Adapter needs a list, so some functions which HashSet doesnt provide
can be used. (Index-based array operations)
Custom implementation of an ArrayList which acts like a HashSet. Theres also
an internal HashSet running in background so that checks for duplicates can be executed in O(1).
List.contain(...) needs O(n).
The tests added should check the behaviour of the custom list, to act like
kind of a HashSet. Operation comparison between custom List and HashSet. Tests are parameterized so checks for every new object that uses this list implementation can be added within 4 lines of code.
HashSet does automatically sort the data after operations. Since we need logic of a programming language, this should be users task and we'll keep insertion order like a list.