CATROID-282 Description not updated/uploaded correctly
When uploading a project, you can change description in the upload dialog. But Changes were made only locally.
The web server doesn't fetch the projectDescription Tag of the HTTPRequest. It fetches the description from the
code.xml, which was saved "after" Upload. So the old description still remained in Upload (code.xml)