CATROID-499 WebRequestBrick to data and testing category
The WebRequestBrick is now a red DataBrick and is removed from the ControlBrickList. It's now present in the DataBrickList and in the TestingBrickList.

CATROID-499 WebRequestBrick to data and testing category

The WebRequestBrick is now a red DataBrick and is removed from the ControlBrickList. It's now present in the DataBrickList and in the TestingBrickList.

changes