CATROID-351 Correct pen brick draws in Maximize mode
If you change from STRETCH mode to MAXIMIZE mode, the draws
of the pen brick are not correctly scaled. So the draws have
the wrong position.
The change of the screen mode is performed in the StageListener.
Proper scaling of the ShapeRenderer (which is used to draw) is
applied there. The scaling factor for x and y is also calculated
there.
Also changed if-query in PenActor to increase performance.