CATROID-623 "if( condition , then , else )" Function
Made the implementation of formulas with more than two parameters possible for this and future formulas by adding an additional list of arbitrary size of children to each formula element. The implementation for formulas with 2 or less children stayed the same and it is therefore possible to load old project files without conversion.
Implemented the if then else function using this new implementation.