Fix the memory leak in ShowBubbleActor
Add close() method to class ShowBubbleActor and dispose
the pixmaps and textures used in the constructor to prevent
memory leaks. They (Pixmap, Texture) are holding native memory
resources if dispose is not called at the end of the lifetime.
Finally the close() method is called when BubbleActor
got removed from BubbleActorList.