CATROID-861 "Broadcast XYZ and wait" does not wait for all "When you receive XYZ" scripts to complete
The bug was caused by using a set with no duplicates of sprites for the waiting. If the brick had to wait
for two scripts in the same sprite it only waited for the first one because of that.
Also removed redundant notifiying in ScriptSequenceActionWithWaiter.kt which caused a double notify because
this is already handled by the ThreadScheduler. Also had to change a testcase because another already existing
issue prevents it functioning correctly.