DEVOPS-491 Include Kotlin classes in coverage report
Fixes a small bug in gradle task. The subdirectory where the class-files
for Kotlin classes are dumped starts with a lower case letter
(camel case). In the gradle task however it was defined capitalized.
For that reason the class-files were not found and therefore not
included in the coverage report.