CATROID-893 Rework Broadcast spinner
Replace the spinner for message selection with an own fragment

Affected bricks are:

- BroadcastAndWaitBrick
- BroadcastReceiveBrick
- BroadcastSendBrick

When clicking on the message on the broadcast brick
there is now this new fragment BroadcastListFragment which
shows all currently used messages. Here all messages are managed.

Previous tests needed to be adopted because they used a spinner
for broadcast message selection, adding and editing.

The BroadcastBrickTest used roboelectric to test adding and renaming
broadcast messages. This behavioral checking is mostly done in espresso tests.
It was not possible to simply fix this test and it does not provide
additional coverage therefore it was deleted.