CATROID-581 Refactor handling of screen sizes in StageActivity
In the StageActivity there exist a few methods which are responsible
for resizing the project's resolution in order the fit the mobile
device's screen.
This code was extracted into a new class (-> Resolution) and tests
were created. Same or similar calculations in other places throughout
the code base were refactored to use this new class in order to
consolidate the responsibility for resizing in one single place.