CATROID-1267 Names of objects in backpacks get changed
When packing and unpacking the object name was previously
chosen with getUniqueNameInNameables of UniqueNameProvider.

This method always appends (1) at the end of the name
even if there is no name-duplicate.

For backpack-operations a new method is used:
getUniqueNameInNameablesIfNecessary

If there is a duplicate name the old method gets called,
otherwise the original name is taken.

affected objects:
- Scene
- Sprite
- Sound
- Look
- Script