CATROID-1403 Fix "write variable to file"-brick resource bug
The target directory for both the "Write variable __ to file __" and
"Read variable from file __" brick is now the shared Download-directory
regardless of the Android version. Since the read-brick only needs to
be able to read files that have previously been created by a write-
brick, we do not need to use the System File Picker and can use
the ordinary way via java.io.File (adding new files to the downloads
collection and reading files that were previously created by the same
app do not require any permissions).
Minor changes along the way: 1) The StageResourceHolder has also
been changed to now consider the Download-directory and print an
appropriate error message if it is not available.
2) both bricks can now work with any extension. If no extension
was supplied, "txt" is used as default.