IDE-29 Change to targetSDKVersion 33
The version of the gradle plugin had to be raised to 7.0.0 and Java
to version 11 in order for the build to work for the new target
Android API 33.
The version of many other dependencies had to be increased as well.
All of them concerning test libraries (Mockito, etc.). Otherwise the
tests suites will not run.
The signed/released aar-files of release 06/2023 are included as well.