Fix Removing Request Install Permissions (#4645)
REQUEST_INSTALL_PACKAGES was used to install standalone APKs. This feature has been disabled due to security concerns, so REQUEST_INSTALL_PACKAGES is not needed anymore.

However, the The REQUEST_INSTALL_PACKAGES permission removal apparently was then ignored. The permission was no longer in our manifest, but if you looked at the apk using Android Studio apk analyzer, it still was there. Apparently it is added in the manifest merge step because of some library we have in it. Figuring out which one it is is not so easy, so we now explicitly state in the manifest that the permission must not be included.

Co-authored-by: Wolfgang Slany <1747888+wslany@users.noreply.github.com>
Co-authored-by: Julia <j.herold@student.tugraz.at>