Updated README.md for Catroid
Expanded on sections of the README.md, including a new link in the references, a more detailed introduction section thats goes into the mission and more context on Catroid and Catrobat. A new Getting started section that allow all levels of developers to have a better footing into contributing to this project.