CATROID-1607 Fix bluetooth for Android API 31 onward
Bluetooth now works again across all devices. Depending on the device
running the app either the new or the legacy bluetooth permissions are
used.
This is achieved by streamlining/simplifying the permission handling
process for all features that require bluetooth permissions.
And a little bit of boyscouting through the code.